requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
transformers
sqlalchemy
pydantic
python-dotenv
pytest
httpx
